# public/openapi.yaml
openapi: 3.1.0
info:
  title: "What's on? API"
  description: |
    Aggregate movie and TV show ratings from IMDb, Rotten Tomatoes, AlloCiné,
    Metacritic, Letterboxd, BetaSeries, SensCritique, TMDB, and Trakt.

    **Rate limiting** — 50 requests per hour and 200 per day. Pass an optional
    `api_key` query parameter for a higher limit. The `Retry-After` header is set
    when the limit is reached (HTTP 429).

    **Ratings scale** — `ratings_average` and `minimum_ratings` use a normalised
    0–5 scale across all sources (IMDb ÷ 2, Metacritic ÷ 20, Rotten Tomatoes ÷ 20,
    AlloCiné native, etc.).
  version: 3.39.5
  contact:
    name: Pierre Vanoverschelde
    url: https://github.com/pierrevano/whatson-api
  license:
    name: See LICENSE
    url: https://github.com/pierrevano/whatson-api/blob/main/LICENSE
servers:
  - url: https://whatson-api.onrender.com
    description: Production
security:
  - apiKey: []

# ---------------------------------------------------------------------------
# Reusable components
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: query
      name: api_key
      description: Optional API key. Requests without a key are subject to the public rate limit.
  parameters:
...
    releaseDateRange:
      name: release_date
      in: query
      description: |
        Date range filter. Either bound is optional.
        Format: `from:YYYY-MM-DD,to:YYYY-MM-DD`
      schema:
        type: string
        example: "from:2020-01-01,to:2024-12-31"
...
    limit:
      name: limit
      in: query
      description: Number of results per page (default 20).
      schema:
        type: integer
        minimum: 1
        maximum: 900
        default: 20
    page:
      name: page
      in: query
      description: Page number, 1-based (default 1).
      schema:
        type: integer
        minimum: 1
        default: 1
  schemas:
...
    RatedEpisode:
      type: object
      properties:
        season:
          type: integer
        episode:
          type: integer
        title:
          type: string
        description:
          type: string
        id:
          type: string
          description: IMDb episode ID (e.g. tt1267295).
        url:
          type: string
          format: uri
        release_date:
          type: string
          format: date
        users_rating:
          type: number
          description: IMDb user rating 0–10.
        users_rating_count:
          type: integer
    AiringEpisode:
      allOf:
        - $ref: "#/components/schemas/RatedEpisode"
        - type: object
          properties:
            episode_type:
              type: string
    SeasonMetadata:
      type: object
      properties:
        season_number:
          type: integer
        episodes_count:
          type: integer
        average_users_rating:
          type: number
          description: Average IMDb user rating for the season.
        users_rating_count:
          type: integer
        highest_episode:
          description: Highest IMDb-rated episode within this season.
          $ref: "#/components/schemas/RatedEpisode"
        lowest_episode:
          description: Lowest IMDb-rated episode within this season.
          $ref: "#/components/schemas/RatedEpisode"
        rating_distribution:
          type: object
          description: Number of episodes grouped by IMDb rating ranges (1–10). Included when `rating_distribution` is in `append_to_response`.
        rating_distribution_episodes:
          type: object
          description: Episodes grouped by IMDb rating ranges (1–10). Included when `rating_distribution_episodes` is in `append_to_response`.
    TvShowSeasons:
      type: object
      properties:
        id:
          type: integer
        item_type:
          type: string
          enum: [tvshow]
        title:
          type: string
        image:
          type: string
          format: uri
        networks:
          type: object
        seasons_number:
          type: integer
        status:
          type: string
          enum: [Canceled, Ended, Ongoing, Pilot, Unknown]
        seasons:
          type: array
          items:
            $ref: "#/components/schemas/SeasonMetadata"
        # Optional episode highlights (append_to_response) — across all seasons
        highest_episode:
          description: Highest IMDb-rated episode across all seasons. Included when `highest_episode` is in `append_to_response`.
          $ref: "#/components/schemas/RatedEpisode"
        lowest_episode:
          description: Lowest IMDb-rated episode across all seasons. Included when `lowest_episode` is in `append_to_response`.
          $ref: "#/components/schemas/RatedEpisode"
        last_episode:
          description: Most recently aired episode. Included when `last_episode` is in `append_to_response`.
          $ref: "#/components/schemas/AiringEpisode"
        next_episode:
          description: Next scheduled episode. Included when `next_episode` is in `append_to_response`.
          $ref: "#/components/schemas/AiringEpisode"
    TvShowSeasonEpisodes:
      type: object
      properties:
        id:
          type: integer
        item_type:
          type: string
          enum: [tvshow]
        title:
          type: string
        image:
          type: string
          format: uri
        networks:
          type: object
        season_number:
          type: integer
        seasons_number:
          type: integer
        status:
          type: string
          enum: [Canceled, Ended, Ongoing, Pilot, Unknown]
        total_episodes:
          type: integer
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/RatedEpisode"
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: "No matching items found."
  responses:
    RateLimitExceeded:
      description: Rate limit exceeded. Retry after the indicated delay.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /updates:
    get:
      operationId: getUpdates
      summary: Get updated title IDs
      description: Returns IDs of titles added or updated since a given date, grouped by item type. Requires a sponsor API key.
      tags: [Titles]
      parameters:
        - name: since
          in: query
          required: true
          description: ISO 8601 date (YYYY-MM-DD) or timestamp with a timezone. The lower bound is inclusive.
          schema:
            type: string
            example: "2026-01-01T00:00:00.000Z"
        - name: item_type
          in: query
          description: One or both item types, separated by a comma.
          schema:
            type: string
            enum: [movie, tvshow, "movie,tvshow", "tvshow,movie"]
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Paginated IDs grouped by item type.
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  updated:
                    type: object
                    properties:
                      movie:
                        type: array
                        items:
                          type: integer
                      tvshow:
                        type: array
                        items:
                          type: integer
                  total_pages:
                    type: integer
                  total_results:
                    type: integer
        "400":
          description: Invalid query parameters.
        "403":
          description: A sponsor API key is required.
        "405":
          description: Method not allowed.
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          description: Internal server error.
        "503":
          description: Service unavailable.
...
  /tvshow/{id}/seasons:
    get:
      operationId: getTvShowSeasons
      summary: Get all seasons for a TV show
      description: |
        Returns per-season episode counts and IMDb rating summaries (average rating,
        highest-rated episode, lowest-rated episode).
      tags: [TV Shows]
      parameters:
        - name: id
          in: path
          required: true
          description: TMDB numeric ID of the TV show.
          schema:
            type: integer
            example: 1396
        - name: append_to_response
          in: query
          description: |
            Optional fields to include. Two scopes:

            **Added to each season object:** `rating_distribution` (episode count per rating band),
            `rating_distribution_episodes` (episode list per rating band).

            **Added to the root response (across all seasons):** `highest_episode`, `lowest_episode`,
            `last_episode`, `next_episode`.
          schema:
            type: string
            example: "rating_distribution,highest_episode"
      responses:
        "200":
          description: TV show seasons summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TvShowSeasons"
        "400":
          description: Invalid query parameters.
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method not allowed.
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          description: Internal server error.
        "503":
          description: Service unavailable.
  /tvshow/{id}/seasons/{season_number}/episodes:
    get:
      operationId: getTvShowSeasonEpisodes
      summary: Get all episodes for a specific season
      description: Returns the list of episodes for a specific season, with optional episode-level filters (minimum rating, release date range).
      tags: [TV Shows]
      parameters:
        - name: id
          in: path
          required: true
          description: TMDB numeric ID of the TV show.
          schema:
            type: integer
            example: 1396
        - name: season_number
          in: path
          required: true
          description: Season number (1-based).
          schema:
            type: integer
            minimum: 1
            example: 3
        - name: minimum_ratings
          in: query
          description: Minimum IMDb user rating (0–10 scale).
          schema:
            type: number
            minimum: 0
            maximum: 10
        - $ref: "#/components/parameters/releaseDateRange"
      responses:
        "200":
          description: Episodes for the requested season.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TvShowSeasonEpisodes"
        "400":
          description: Invalid query parameters.
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method not allowed.
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          description: Internal server error.
        "503":
          description: Service unavailable.
...